pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docaroapp"
version = "0.1.0"
description = "Modern Document Pipeline"
authors = [{name = "Gregorfun", email = "gregor@example.com"}]
dependencies = [
    "flask",
    "gunicorn",
    "argon2-cffi",
    "rq",
    "redis",
    "pdf2image",
    "pytesseract",
    "pillow",
    "python-dateutil",
    "PyPDF2",
    "pdfplumber",
    "PyYAML"
]
requires-python = ">=3.9"

[tool.pytest.ini_options]
testpaths = ["tests", "core"]
python_files = "test_*.py"
pythonpath = ["."]
addopts = "-v --cov=app --cov=core"
markers = [
    "benchmark: marks performance benchmark tests (deselect with '-m \"not benchmark\"')",
]

[tool.black]
line-length = 120
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 120

[tool.ruff]
line-length = 120
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W"]
ignore = ["E501"]
